requirements.txt

fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-community
langchain-groq
langchain-huggingface
unstructured
pymupdf==1.21.0  # fitz (PyMuPDF)
faiss-cpu
python-multipart
boto3
fitz